+++ .gitignore
@@ -77,3 +77,4 @@ dist_cache
 tsconfig.tsbuildinfo
 ingest-attachment-*.zip
 tsdoc-metadata.json
+pods/front/dist
