--- server/server/kube/ingress.yml
@@ -1,26 +1,0 @@
-
-apiVersion: networking.k8s.io/v1
-kind: Ingress
-metadata:
-  annotations:
-    kubernetes.io/ingress.class: "nginx"
-    cert-manager.io/issuer: "letsencrypt-prod"
-    nginx.org/proxy-read-timeout: "3h"
-    nginx.org/proxy-send-timeout: "3h"
-  name: transactor-ingress
-spec:
-  tls:
-  - hosts:
-    - transactor.hc.engineering
-    secretName: transactor-tls
-  rules:
-  - host: transactor.hc.engineering
-    http:
-      paths:
-      - backend:
-          service:
-            name: transactor
-            port: 
-              number: 80
-        path: /
-        pathType: Prefix
