--- server/server/kube/transactor.yml
@@ -1,60 +1,0 @@
-
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: transactor
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: transactor
-  template:
-    metadata:
-      labels:
-        app: transactor
-    spec:
-      containers:
-        - name: app
-          image: hardcoreeng/transactor
-          ports:
-            - containerPort: 3333
-          imagePullPolicy: Always
-          env:
-          - name: FRONT_URL
-            value: https://front.hc.engineering/
-          - name: MONGO_URL
-            valueFrom:
-              secretKeyRef:
-                name: mongodb
-                key: url
-          - name: ELASTIC_URL
-            valueFrom:
-              secretKeyRef:
-                name: elastic
-                key: url
-          - name: MINIO_ENDPOINT
-            valueFrom:
-              secretKeyRef:
-                name: minio
-                key: endpoint
-          - name: MINIO_ACCESS_KEY
-            valueFrom:
-              secretKeyRef:
-                name: minio
-                key: accessKey
-          - name: MINIO_SECRET_KEY
-            valueFrom:
-              secretKeyRef:
-                name: minio
-                key: secretKey
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: transactor
-spec:
-  selector:
-    app: transactor
-  ports:
-    - port: 80
-      targetPort: 3333
