+++ server/server/package.json
@@ -5,13 +5,8 @@
   "author": "Anticrm Platform Contributors",
   "license": "EPL-2.0",
   "scripts": {
-    "start": "cross-env MONGO_URL=mongodb://localhost:27017 ELASTIC_URL=http://localhost:9200 FRONT_URL=http://localhost:8087 MINIO_ENDPOINT=localhost MINIO_ACCESS_KEY=minioadmin MINIO_SECRET_KEY=minioadmin METRICS_CONSOLE=true SERVER_SECRET=secret ts-node src/__start.ts",
     "build": "heft build",
     "lint:fix": "eslint --fix src",
-    "bundle": "esbuild src/__start.ts --bundle --platform=node > bundle.js",
-    "docker:build": "docker build -t hardcoreeng/transactor .",
-    "docker:staging": "../../common/scripts/docker_tag.sh hardcoreeng/transactor staging",
-    "docker:push": "../../common/scripts/docker_tag.sh hardcoreeng/transactor",
     "build:watch": "tsc",
     "lint": "eslint src",
     "format": "prettier --write src && eslint --fix src"
@@ -28,7 +23,6 @@
     "eslint": "^8.26.0",
     "@types/ws": "^8.5.3",
     "ts-node": "^10.8.0",
-    "esbuild": "^0.12.26",
     "@typescript-eslint/parser": "^5.41.0",
     "eslint-config-standard-with-typescript": "^23.0.0",
     "prettier": "^2.7.1",
@@ -41,41 +35,10 @@
     "@hcengineering/platform": "^0.6.7",
     "@hcengineering/server-core": "~0.6.1",
     "@hcengineering/server-ws": "~0.6.11",
-    "@hcengineering/server-attachment": "~0.6.1",
-    "@hcengineering/server-attachment-resources": "~0.6.0",
     "@hcengineering/mongo": "~0.6.1",
     "@hcengineering/elastic": "~0.6.0",
     "elastic-apm-node": "~3.26.0",
-    "minio": "^7.0.26",
-    "@hcengineering/server-contact": "~0.6.1",
-    "@hcengineering/server-contact-resources": "~0.6.0",
-    "@hcengineering/server-notification": "^0.6.0",
-    "@hcengineering/server-notification-resources": "~0.6.0",
-    "@hcengineering/server-setting": "~0.6.0",
-    "@hcengineering/server-setting-resources": "~0.6.0",
-    "@hcengineering/server-chunter": "~0.6.0",
-    "@hcengineering/server-chunter-resources": "~0.6.0",
-    "@hcengineering/server-inventory": "~0.6.0",
-    "@hcengineering/server-inventory-resources": "~0.6.0",
-    "@hcengineering/server-lead": "~0.6.0",
-    "@hcengineering/server-lead-resources": "~0.6.0",
-    "@hcengineering/server-recruit": "~0.6.0",
-    "@hcengineering/server-recruit-resources": "~0.6.0",
-    "@hcengineering/server-task": "~0.6.0",
-    "@hcengineering/server-task-resources": "~0.6.0",
-    "@hcengineering/server-tracker": "~0.6.0",
-    "@hcengineering/server-tracker-resources": "~0.6.0",
-    "@hcengineering/server-tags": "~0.6.0",
-    "@hcengineering/server-tags-resources": "~0.6.0",
-    "@hcengineering/server-calendar": "~0.6.0",
-    "@hcengineering/server-calendar-resources": "~0.6.0",
-    "@hcengineering/server-gmail": "~0.6.0",
-    "@hcengineering/server-gmail-resources": "~0.6.0",
-    "@hcengineering/server-preference": "~0.6.0",
-    "@hcengineering/server-telegram": "~0.6.0",
-    "@hcengineering/server-telegram-resources": "~0.6.0",
-    "@hcengineering/server-hr": "~0.6.0",
-    "@hcengineering/server-hr-resources": "~0.6.0",
+    "minio": "^7.0.26",   
     "@hcengineering/server-token": "~0.6.0",
     "@hcengineering/middleware": "~0.6.0"
   }
