+++ tools/apm/package.json
@@ -1,11 +1,12 @@
 {
-  "name": "@hcengineering/server-tool",
+  "name": "@hcengineering/apm",
   "version": "0.6.0",
   "main": "lib/index.js",
   "author": "Anticrm Platform Contributors",
   "license": "EPL-2.0",
+  "bin": "./lib/apm.js",
   "scripts": {
-    "build": "heft build",
+    "build": "heft build && esbuild src/index.ts --bundle --minify --platform=node > ./apm.js && echo 'build'",
     "build:watch": "tsc",
     "lint:fix": "eslint --fix src",
     "lint": "eslint src",
@@ -24,19 +25,11 @@
     "prettier": "^2.7.1",
     "@rushstack/heft": "^0.47.9",
     "typescript": "^4.3.5",
-    "@types/minio": "~7.0.11",
-    "@types/ws": "^8.5.3"
+    "esbuild": "^0.15.13",
+    "@types/node": "~16.11.12"
   },
   "dependencies": {
-    "mongodb": "^4.9.0",
-    "@hcengineering/platform": "^0.6.7",    
-    "minio": "^7.0.26",
-    "@hcengineering/core": "^0.6.17",
-    "@hcengineering/contact": "~0.6.5",
-    "@hcengineering/client-resources": "~0.6.4",
-    "@hcengineering/client": "^0.6.3",
-    "ws": "^8.10.0",
-    "@hcengineering/model": "~0.6.0",
-    "@hcengineering/server-token": "~0.6.0"
+    "comment-json": "^4.2.2",
+    "commander": "^8.1.0"
   }
 }
