+++ tools/apm/src/template.ts
@@ -1,0 +1,1 @@
+export async function createTemplate (root: string, platformRoot: string): Promise<void> {}
