+++ tools/apm/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "./node_modules/@hcengineering/platform-rig/profiles/default/tsconfig.json",
+
+  "compilerOptions": {
+    "rootDir": "./src",
+    "outDir": "./lib",
+    "types": ["node"],
+    "esModuleInterop": true
+  }
+}
